Cleaning Up Old Code
In preparation for some major additions to the website tomorrow I took
the liberty of cleaning up some of the code. I've also removed the
paragraph on the homepage and replaced it with a single line of dummy
text, do not be alarmed, this is just to rectify some structure issues.